New Updated

+++ camera/views.py
@@ -42,72 +42,60 @@ def success():
 
 
 @login_required
+@bp.route("/images/image_actions/", methods=["POST"])
+def image_actions():
+    data = request.json
+    
+    selected_action = data.get("action")
+    selected_images = data.get("images")
+
+    if selected_action == "delete":
+        for id in selected_images:
+            image = Image.query.get(int(id))
+            
+            print(image)
+
+            try:
+                db.session.delete(image)
+                db.session.commit()
+
+            except Exception as e:
+                flash("Something went wrong.", "error")
+
+    elif selected_action == "approve":
+        for id in selected_images:
+            image = Image.query.get(int(id))
+            
+            print(image)
+
+            image.approved = True
+
+            try:
+                db.session.commit()
+
+            except Exception as e:
+                flash("Something went wrong.", "error")
+
+    elif selected_action == "disapprove":
+        for id in selected_images:
+            image = Image.query.get(int(id))
+            
+            print(image)
+
+            image.approved = False
+
+            try:
+                db.session.commit()
+
+            except Exception as e:
+                flash("Something went wrong.", "error")
+
+    return jsonify({"message": "Action performed successfully"})
+
+
+@login_required
 @bp.route("/images/", methods=["GET"])
 def images():
     all_images = Image.query.all()
 
     return render_template("images.html", images=all_images)
-
-
-@bp.route("/images/delete/<id>/", methods=["POST"])
-def delete(id):
-    image = Image.query.get(int(id))
-
-    print(image)
-
-    try:
-        db.session.delete(image)
-
-    except Exception as e:
-        flash("Something went wrong.", "error")
-
-        return redirect(url_for("home.index"))
-
-    else:
-        db.session.commit()
-
-        return redirect(url_for("home.index"))
-
-
-@bp.route("/images/approve/<id>/", methods=["POST"])
-def approve(id):
-    image = Image.query.get(int(id))
-
-    print(image)
-
-    image.approved = True
-
-    try:
-        db.session.commit()
-
-    except Exception as e:
-        flash("Something went wrong.", "error")
-
-        return redirect(url_for("home.index"))
-
-    else:
-        db.session.commit()
-
-        return redirect(url_for("home.index"))
-
-
-@bp.route("/images/disapprove/<id>/", methods=["POST"])
-def disapprove(id):
-    image = Image.query.get(int(id))
-
-    print(image)
-
-    image.approved = False
-
-    try:
-        db.session.commit()
-
-    except Exception as e:
-        flash("Something went wrong.", "error")
-
-        return redirect(url_for("home.index"))
-
-    else:
-        db.session.commit()
-
-        return redirect(url_for("home.index"))
